backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
pydantic>=2.11.0
pydantic-settings>=2.10.0
httpx>=0.28.0
supabase>=2.15.0
langgraph>=0.2.65
langchain-core>=0.3.68
langchain-openai>=0.2.14
python-multipart>=0.0.20
pypdf>=5.4.0
python-docx>=1.1.2
python-pptx>=1.0.2
pillow>=11.3.0
pytesseract>=0.3.13
